chore: dotnet-guidelines Tier B2 — member ordering + type co-location (pure moves)
Member order (constants, fields, properties, public methods, private methods):
MessagesBackfillJob, ReactionsBackfillJob, ResolvedFks records, OpenRouterClient,
AttachmentUrlRefreshService, MemberEventHandler, StatsController (helpers below
actions, VoiceSegmentMinutes with the consts), TableMeta, RawEventSerializationTests.

Co-location: ColumnKind + TableMeta out of SchemaCatalog.cs;
NullableSnowflakeJsonConverter + DashboardJson out of SnowflakeJsonConverter.cs;
UuidV7Extensions out of DiscordDbContext.cs. No code changes, moves only.
