+++ src/DiscordEventService/Data/DiscordDbContext.cs
@@ -145,16 +145,3 @@ internal interface ITimestamped
     DateTime FirstSeenUtc { get; set; }
     DateTime LastUpdatedUtc { get; set; }
 }
-
-internal static class UuidV7Extensions
-{
-    public static void ConfigureUuidGeneration(this ModelBuilder modelBuilder)
-    {
-        foreach (var entityType in modelBuilder.Model.GetEntityTypes())
-        {
-            var idProperty = entityType.FindProperty("Id");
-            if (idProperty is not null && idProperty.ClrType == typeof(Guid))
-                idProperty.SetDefaultValueSql("uuidv7()");
-        }
-    }
-}
