+++ src/DiscordEventService/Services/Pipeline/ResolvedFks.cs
@@ -2,15 +2,15 @@ namespace DiscordEventService.Services.Pipeline;
 
 internal sealed record ResolvedFks(bool Success, Guid GuildId, Guid ChannelId, Guid UserId)
 {
+    public static readonly ResolvedFks Failed = new ResolvedFks(false, default, default, default);
+
     public static ResolvedFks Resolved(Guid guildId, Guid channelId, Guid userId) =>
         new ResolvedFks(true, guildId, channelId, userId);
-
-    public static readonly ResolvedFks Failed = new ResolvedFks(false, default, default, default);
 }
 
 internal sealed record ResolvedUserFks(bool Success, Guid GuildId, Guid UserId)
 {
-    public static ResolvedUserFks Resolved(Guid guildId, Guid userId) => new ResolvedUserFks(true, guildId, userId);
-
     public static readonly ResolvedUserFks Failed = new ResolvedUserFks(false, default, default);
+
+    public static ResolvedUserFks Resolved(Guid guildId, Guid userId) => new ResolvedUserFks(true, guildId, userId);
 }
